split JsonSErviceOpts and JsonService
